+++ sistema_academia/src/main/java/com/example/sistema_academia/infrastructure/adapters/TreinoJPARep.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.sistema_academia.domain.entities.Treino;
 
-public interface TreinoJPARep extends JpaRepository<Treino, Integer> {
+public interface TreinoJPARep extends JpaRepository<Treino, Long> {
     
     List<Treino> findByCpfAluno(String cpfAluno);
 }
